% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-objects.R
\docType{data}
\name{kawaii_palettes}
\alias{kawaii_palettes}
\alias{kawaii_bi_01}
\alias{kawaii_tri_01}
\alias{kawaii_tri_02}
\alias{kawaii_tri_03}
\alias{kawaii_tri_04}
\alias{kawaii_tri_05}
\alias{kawaii_tri_06}
\alias{kawaii_tri_07}
\alias{kawaii_tri_08}
\alias{kawaii_tri_09}
\alias{kawaii_tri_10}
\alias{kawaii_tri_11}
\alias{kawaii_tri_12}
\alias{kawaii_tri_13}
\alias{kawaii_tri_14}
\alias{kawaii_tri_15}
\alias{kawaii_tri_16}
\alias{kawaii_tri_17}
\alias{kawaii_tri_18}
\alias{kawaii_tri_19}
\alias{kawaii_four_01}
\alias{kawaii_four_02}
\alias{kawaii_multi_01}
\alias{kawaii_multi_02}
\alias{kawaii_multi_03}
\alias{kawaii_multi_04}
\alias{kawaii_multi_05}
\alias{kawaii_multi_06}
\alias{kawaii_multi_07}
\alias{kawaii_multi_08}
\title{Kawaii palettes}
\format{
A character vector of hex colour strings.

\code{kawaii_bi_01}: bicolor, 2 colours (#5C4D8C, #D0D959).

\code{kawaii_tri_01}: tricolor, 3 colours (#F4C8DD, #0097DC, #3BB699).

\code{kawaii_tri_02}: tricolor, 3 colours (#4AB0DF, #D84D96, #8255A0).

\code{kawaii_tri_03}: tricolor, 3 colours (#E7BB7F, #E5C2DB, #EADFCA).

\code{kawaii_tri_04}: tricolor, 3 colours (#FFF462, #AFD9B9, #AFC0E3).

\code{kawaii_tri_05}: tricolor, 3 colours (#DE1628, #27B28D, #F2CB4F).

\code{kawaii_tri_06}: tricolor, 3 colours (#207EB8, #2D7C3A, #EEDD28).

\code{kawaii_tri_07}: tricolor, 3 colours (#3D7EB4, #DF4472, #E68C5F).

\code{kawaii_tri_08}: tricolor, 3 colours (#FADBDA, #AFD9B9, #FFFCD1).

\code{kawaii_tri_09}: tricolor, 3 colours (#FFF462, #FADBDA, #7ECEF4).

\code{kawaii_tri_10}: tricolor, 3 colours (#FADBDA, #C8E7E9, #AFC0E3).

\code{kawaii_tri_11}: tricolor, 3 colours (#FCEAD4, #470B00, #FDD700).

\code{kawaii_tri_12}: tricolor, 3 colours (#F7E2E7, #500047, #FDD700).

\code{kawaii_tri_13}: tricolor, 3 colours (#F9F21E, #B6977B, #A488D3).

\code{kawaii_tri_14}: tricolor, 3 colours (#1D967A, #FDD35C, #E94D3B).

\code{kawaii_tri_15}: tricolor, 3 colours (#FFF462, #3AB483, #AFC0E3).

\code{kawaii_tri_16}: tricolor, 3 colours (#EEF4E7, #003C0B, #FDD700).

\code{kawaii_tri_17}: tricolor, 3 colours (#E1F1F7, #171776, #FDD700).

\code{kawaii_tri_18}: tricolor, 3 colours (#A6F1F7, #EA8A71, #51E7C4).

\code{kawaii_tri_19}: tricolor, 3 colours (#E7242E, #006356, #EFC757). Contains a provisional reading; see \code{palettes_needing_review()}.

\code{kawaii_four_01}: four-color, 4 colours (#CFC83B, #E79970, #9BCCD0, #2E8D3A).

\code{kawaii_four_02}: four-color, 4 colours (#79792D, #75B8BC, #D46932, #3E2C27).

\code{kawaii_multi_01}: multicolor, 5 colours (#F4C8DD, #0097DC, #3BB699, #FCCE60, #E94740).

\code{kawaii_multi_02}: multicolor, 6 colours (#F6AA96, #E9470C, #8EAC7A, #77BD5B, #0091D4, #F0C82B). Contains a provisional reading; see \code{palettes_needing_review()}.

\code{kawaii_multi_03}: multicolor, 7 colours (#C24831, #25A64C, #F7A211, #FF9EB2, #3D308A, #B26054, #84D2C2).

\code{kawaii_multi_04}: multicolor, 5 colours (#EECF24, #DC421F, #305761, #DDBA8F, #E3A52A).

\code{kawaii_multi_05}: multicolor, 5 colours (#E9BF35, #AA95E0, #A4AC49, #DC7B8F, #7695AA).

\code{kawaii_multi_06}: multicolor, 5 colours (#246488, #51662E, #1B2D45, #E8D3CA, #C4BFC6).

\code{kawaii_multi_07}: multicolor, 5 colours (#735944, #918D8D, #D6AF49, #BD1A22, #546F32).

\code{kawaii_multi_08}: multicolor, 5 colours (#D6A31B, #EA5034, #3B65B0, #56B552, #F4DF2A).
}
\usage{
kawaii_bi_01

kawaii_tri_01

kawaii_tri_02

kawaii_tri_03

kawaii_tri_04

kawaii_tri_05

kawaii_tri_06

kawaii_tri_07

kawaii_tri_08

kawaii_tri_09

kawaii_tri_10

kawaii_tri_11

kawaii_tri_12

kawaii_tri_13

kawaii_tri_14

kawaii_tri_15

kawaii_tri_16

kawaii_tri_17

kawaii_tri_18

kawaii_tri_19

kawaii_four_01

kawaii_four_02

kawaii_multi_01

kawaii_multi_02

kawaii_multi_03

kawaii_multi_04

kawaii_multi_05

kawaii_multi_06

kawaii_multi_07

kawaii_multi_08
}
\description{
The 30 palettes of the Kawaii collection, each a character vector of
hex colours in printed swatch order.
}
\details{
2 of them contain a provisional reading and are marked
\code{"review"} in the catalogue; see \code{\link[=palettes_needing_review]{palettes_needing_review()}}.
}
\examples{
kawaii_bi_01
show_palette(kawaii_bi_01)
}
\seealso{
\code{\link[=palettes]{palettes()}}, \code{\link[=show_palette]{show_palette()}}, \code{\link[=scale_colour_japanesecolors]{scale_colour_japanesecolors()}}
}
\keyword{datasets}
